+++ src/HeadlessH5PServiceProvider.php
@@ -3,12 +3,15 @@
 namespace EscolaLms\HeadlessH5P;
 
 use Illuminate\Support\ServiceProvider;
+use EscolaLms\HeadlessH5P\HeadlessH5P;
 
 class HeadlessH5PServiceProvider extends ServiceProvider
 {
     public function register(): void
     {
-        // $this->injectContract(self::CONTRACTS);
+        $this->app->singleton(H5PCore::class, function ($app) {
+            return new HeadlessH5P();
+        });
     }
 
     public function boot(): void
@@ -16,6 +19,6 @@ class HeadlessH5PServiceProvider extends ServiceProvider
         $this->loadRoutesFrom(__DIR__ . '/routes.php');
         //$this->loadFactoriesFrom(__DIR__ . '/../database/factories');
         $this->loadMigrationsFrom(__DIR__ . '/../database/migrations');
-        //$this->loadTranslationsFrom(__DIR__ . '/../resources/lang', 'mad');
+        // Load configs
     }
 }
